+++ synch/exceptions.py
@@ -1,0 +1,6 @@
+class SynchException(Exception):
+    """Base Exception """
+
+
+class ConfigurationError(Exception):
+    """raise when config error"""