+++ synch/writer/collapsing_merge_tree.py
@@ -1,5 +1,6 @@
 from typing import Dict
 
+from synch.common import cluster_sql
 from synch.enums import ClickHouseEngine
 from synch.reader import Reader
 from synch.writer.merge_tree import ClickHouseMergeTree
@@ -18,6 +19,9 @@ class ClickHouseCollapsingMergeTree(ClickHouseMergeTree):
         engine_settings: str = None,
         sign_column: str = None,
     ):
+        super(ClickHouseCollapsingMergeTree, self).get_table_create_sql(
+            reader, schema, table, pk, partition_by, engine_settings, sign_column=sign_column
+        )
         select_sql = reader.get_source_select_sql(schema, table, sign_column)
         partition_by_str = ""
         engine_settings_str = ""
@@ -25,7 +29,7 @@ class ClickHouseCollapsingMergeTree(ClickHouseMergeTree):
             partition_by_str = f" PARTITION BY {partition_by} "
         if engine_settings:
             engine_settings_str = f" SETTINGS {engine_settings} "
-        return f"CREATE TABLE {schema}.{table} ENGINE = {self.engine}({sign_column}) {partition_by_str} ORDER BY {pk} {engine_settings_str} AS {select_sql} limit 0"
+        return f"CREATE TABLE {schema}.{table}{cluster_sql(self.cluster_name)}ENGINE = {self.engine}({sign_column}) {partition_by_str} ORDER BY {pk} {engine_settings_str} AS {select_sql} limit 0"
 
     def get_full_insert_sql(self, reader: Reader, schema: str, table: str, sign_column: str = None):
         return f"insert into {schema}.{table} {reader.get_source_select_sql(schema, table, sign_column)}"